Use BigInteger for match/event timestamp columns
On Postgres, SQLAlchemy Integer maps to int4 (32-bit). TBA returns a ~1900
placeholder Unix timestamp (-2208988800) for matches/events with an unknown
time, which underflows int32 and aborts historical builds at 2006 with
psycopg2.errors.NumericValueOutOfRange. CockroachDB INT is 64-bit, so this
never surfaced on the production database. Widen the four timestamp columns
(match.time, match.predicted_time, event.time, team_event.time) to BigInteger;
this matches CockroachDB's width and is future-proof past the 2038 int32 limit.

+++ backend/src/db/models/team_event.py
@@ -1,6 +1,6 @@
 from typing import Any, Dict, Tuple
 
-from sqlalchemy import Boolean, Enum, Float, Integer, String
+from sqlalchemy import BigInteger, Boolean, Enum, Float, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
 from sqlalchemy.sql.schema import ForeignKeyConstraint, PrimaryKeyConstraint
 
@@ -27,7 +27,7 @@ class TeamEventORM(Base, ModelORM):
     ForeignKeyConstraint(["event"], ["events.key"])
 
     """GENERAL"""
-    time: MI = mapped_column(Integer)
+    time: MI = mapped_column(BigInteger)  # Unix timestamp; see match.py note
 
     """API COMPLETENESS"""
     team_name: MS = mapped_column(String(100))
